+++ lib/generators/spectre/templates/spectre_initializer.rb
@@ -3,7 +3,7 @@
 require 'spectre'
 
 Spectre.setup do |config|
-  # Chose your LLM (openai, ollama)
+  # Chose your LLM (openai, ollama, claude, gemini)
   config.default_llm_provider = :openai
 
   config.openai do |openai|
@@ -13,5 +13,13 @@ Spectre.setup do |config|
   config.ollama do |ollama|
     ollama.host = ENV['OLLAMA_HOST']
     ollama.api_key = ENV['OLLAMA_API_KEY']
+  end
+
+  config.claude do |claude|
+    claude.api_key = ENV['ANTHROPIC_API_KEY']
+  end
+
+  config.gemini do |gemini|
+    gemini.api_key = ENV['GEMINI_API_KEY']
   end
 end